param sp analyzer dev

+++ AnalysisAndSuch/ParamSpAnalyzer.py
@@ -41,7 +41,7 @@ def main():
     clear_writeFile()
     print(geffInfo['bounds'][0]*1000, (geffInfo['bounds'][1]+ geffInfo['delta'])*1000, geffInfo['delta']*1000, sep='\n')
     Lcols = [l for l in range(LambdaInfo['bounds'][0], LambdaInfo['bounds'][1]+ LambdaInfo['delta'], LambdaInfo['delta'])]
-    gRows = [g/1000 for g in range(geffInfo['bounds'][0]*1000, (geffInfo['bounds'][1]+ geffInfo['delta'])*1000, geffInfo['delta']*1000)]
+    gRows = [g/1000 for g in range(geffInfo['bounds'][0]*1000, (geffInfo['bounds'][1]+ geffInfo['delta'])*1000, int(geffInfo['delta']*1000))]
     # analyzeThis('LNVF_1000_170')
     print(Lcols)
     print(gRows)
